Refactor logic common to collapsible and non-collapsible Flashbar into separate hook
