+++ src/flashbar/stackable-flashbar.tsx
@@ -1,19 +1,14 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import clsx from 'clsx';
-import React, { useCallback, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useLayoutEffect, useRef, useState } from 'react';
 import customCssProps from '../internal/generated/custom-css-properties';
 import { Flash, focusFlashById } from './flash';
 import { FlashbarProps, StackedFlashbarProps } from './interfaces';
-import { getBaseProps } from '../internal/base-component';
 import InternalIcon from '../icon/internal';
 import { TransitionGroup } from 'react-transition-group';
 import { Transition } from '../internal/components/transition';
-import useBaseComponent from '../internal/hooks/use-base-component';
-import { useContainerBreakpoints } from '../internal/hooks/container-queries';
 import useFocusVisible from '../internal/hooks/focus-visible';
-import { useMergeRefs } from '../internal/hooks/use-merge-refs';
-import { useReducedMotion, useVisualRefresh } from '../internal/hooks/use-visual-mode';
 import { getVisualContextClassname } from '../internal/components/visual-context';
 
 import styles from './styles.css.js';
@@ -24,21 +19,18 @@ import { IconProps } from '../icon/interfaces';
 import { sendToggleMetric } from './internal/analytics';
 
 import Name = IconProps.Name;
+import { useFlashbar } from './common';
 
 export { FlashbarProps };
 
 const maxUnstackedItems = 1;
 
 export default function StackableFlashbar({ items, ...restProps }: FlashbarProps & StackedFlashbarProps) {
-  const { __internalRootRef } = useBaseComponent('Flashbar');
-  const baseProps = getBaseProps(restProps);
-
-  const ref = useRef<HTMLDivElement | null>(null);
-  const [breakpoint, breakpointRef] = useContainerBreakpoints(['xs']);
-  const mergedRef = useMergeRefs(ref, breakpointRef, __internalRootRef);
-
+  const { allItemsHaveId, baseProps, breakpoint, isReducedMotion, isVisualRefresh, mergedRef, ref } = useFlashbar({
+    items,
+    ...restProps,
+  });
   const isFocusVisible = useFocusVisible();
-  const isVisualRefresh = useVisualRefresh();
 
   const [previousItems, setPreviousItems] = useState<ReadonlyArray<FlashbarProps.MessageDefinition>>(items);
   const [nextFocusId, setNextFocusId] = useState<string | null>(null);
@@ -53,9 +45,6 @@ export default function StackableFlashbar({ items, ...restProps }: FlashbarProps
   const [transitioning, setTransitioning] = useState(false);
   const [isFlashbarStackExpanded, setIsFlashbarStackExpanded] = useState(false);
   const [listTop, setListTop] = useState<number>();
-
-  const isReducedMotion = useReducedMotion(breakpointRef as any);
-  const allItemsHaveId = useMemo(() => items.every(item => 'id' in item), [items]);
 
   const flashbarElementId = useUniqueId('flashbar');
   const statusElementId = useUniqueId('toggle');
@@ -100,7 +89,7 @@ export default function StackableFlashbar({ items, ...restProps }: FlashbarProps
     if (nextFocusId) {
       focusFlashById(ref.current, nextFocusId);
     }
-  }, [nextFocusId]);
+  }, [nextFocusId, ref]);
 
   useEffect(() => {
     if (items.length <= maxUnstackedItems) {
